Extract mobile dashboard view
